\name{processed.rt.data}
\alias{processed.rt.data}
\docType{data}
\title{
Processed rt.data
}
\description{
Processed rt.data
}
\usage{data(processed.rt.data)}
\format{
  A data frame with 12 observations on the following 4 variables.
  \describe{
    \item{\code{Subject}}{a numeric vector}
    \item{\code{Block}}{a numeric vector}
    \item{\code{Trial}}{a numeric vector}
    \item{\code{CumMeanRT}}{a numeric vector}
  }
}
\examples{
data(processed.rt.data)
}
\keyword{datasets}
